YOLO_Cat/requirement.txt
numpy
matplotlib>=3.3.0
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
psutil
pandas>=1.1.4
seaborn>=0.11.0
transformers